requirements.txt
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.9.0
greenlet==3.2.3
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
playwright==1.54.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
PyJWT==2.10.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
reportlab==4.4.3
scikit-image==0.25.2
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
SQLAlchemy==2.0.43
sympy==1.14.0
tifffile==2025.8.28
torch==2.8.0
torchvision==0.23.0
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
